Tests/LibWeb/Text/expected/htmlscript-src-reflection.txt

is_string = true
has_protocol = true
ends_with_dummy = true
